requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
cryptography==37.0.4
discord==2.0.0
discord.py==2.0.1
distlib==0.3.4
dnspython==2.2.1
filelock==3.6.0
frozenlist==1.3.1
h11==0.13.0
idna==3.4
iniconfig==1.1.1
multidict==6.0.2
my-modules==1.0
numpy==1.23.3
outcome==1.2.0
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pydantic==1.10.2
pygame==2.1.2
pymongo==4.2.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytube==12.1.0
pytz==2022.2.1
requests==2.28.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.14.1
webdriver-manager==3.8.2
wsproto==1.1.0
yarl==1.8.1
